UserContext error finally fixed. Jesus
