+++ app/_layout.tsx
@@ -5,6 +5,7 @@ import * as SplashScreen from 'expo-splash-screen';
 import { useEffect } from 'react';
 import 'react-native-reanimated';
 
+import { UserContextProvider } from '@/context/user';
 import { useColorScheme } from '@/hooks/useColorScheme';
 
 // Prevent the splash screen from auto-hiding before asset loading is complete.
@@ -27,6 +28,7 @@ export default function RootLayout() {
   }
 
   return (
+    <UserContextProvider>
     <ThemeProvider value={colorScheme === 'dark' ? DarkTheme : DefaultTheme}>
       <Stack>
         <Stack.Screen name="index" options={{ headerShown: false }} />
@@ -36,5 +38,6 @@ export default function RootLayout() {
         <Stack.Screen name="community" />
       </Stack>
     </ThemeProvider>
+    </UserContextProvider>
   );
 }
