+++ context/user.tsx
@@ -1,0 +1,32 @@
+import { ReactNode, Dispatch, createContext, useState, SetStateAction, useEffect } from "react";
+
+type User = {
+  id: number;
+  firstName: string;
+  lastName: string;
+  email: string;
+};
+
+type UserContextType = {
+  user: User | null;
+  setUser: Dispatch<SetStateAction<User | null>>;
+};
+
+export const UserContext = createContext<UserContextType | null>(null);
+
+export const UserContextProvider = ({ children }: { children: ReactNode }) => {
+  const [user, setUser] = useState<User | null>(null);
+
+  useEffect(() => {
+    const user = localStorage.getItem("user");
+    if (user) {
+      setUser(JSON.parse(user));
+    }
+  }, []);
+
+  return (
+    <UserContext.Provider value={{ user, setUser }}>
+      {children}
+    </UserContext.Provider>
+  );
+};
